+++ tests/precision_contradictory_range_condition_finding.adb
@@ -1,0 +1,6 @@
+procedure Precision_Contradictory_Range_Condition_Finding (X : Integer) is
+begin
+   if X > 10 and then X < 5 then
+      null;
+   end if;
+end Precision_Contradictory_Range_Condition_Finding;