+++ tests/precision_redundant_abs_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Redundant_Abs_Clean (X : Integer) is
+   Y : Integer;
+begin
+   Y := abs (X);
+end Precision_Redundant_Abs_Clean;