+++ tests/precision_redundant_unary_minus_clean.adb
@@ -1,0 +1,5 @@
+procedure Precision_Redundant_Unary_Minus_Clean (X : Integer) is
+   Y : Integer;
+begin
+   Y := -X;
+end Precision_Redundant_Unary_Minus_Clean;